+++ fbs_license_manager/migrations/0001_initial.py
@@ -182,3 +182,7 @@ class Migration(migrations.Migration):
             },
         ),
     ]
+
+
+
+